Moved files up in the directory hierarchy, changed the models to more accurately reflect the JSONs...
