+++ database.js
@@ -1,4 +1,5 @@
 const mongoose = require('mongoose');
+
 let mongooseConnectionConfig = { useNewUrlParser: true, useUnifiedTopology: true }
 
 let connectionString = ""
