+++ src/mangaMeat/models/character.js
@@ -1,12 +1,12 @@
 let mongoose = require('mongoose')
 
 let characterSchema = new mongoose.Schema({
-    id: Number,
-    name: String,
-    animeOrigin: String,
-    school: String,
-    age: Number,
-    picture: String
+    ID: Number,
+    Name: String,
+    AnimeOrigin: String,
+    School: String,
+    Age: Number,
+    Picture: String
 })
 
 let characterModel = mongoose.model("Characters", characterSchema)
